backend-fastapi/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
psycopg[binary]==3.3.2
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
pydantic==2.12.5
